Create comparestrings
In Solidity, there is no native function to compare the strings. It is not possible to compare the equality of the two strings. There is a way to do this by checking byte-by-byte data, but it would be a costly operation in terms of gas consumption. Solidity is not ideal for this, so you should think carefully when using this.